Add files via upload
For my final year project, I developed a web-based online voting system with the primary goal of offering college and university students a secure, efficient, and user-friendly means to cast their votes for class representatives. The traditional manual voting system used in universities and colleges has been facing numerous problems such as low voter turnout, absenteeism, lack of interest, and trust issues.

The proposed online voting system aimed to combat these issues by allowing students to participate in elections from the comfort of their own devices, anytime and from anywhere. This approach would effectively reduce absenteeism and increase voter participation, fostering a more inclusive and representative democratic process within academic institutions.

To bring this project to life, I used a combination of web technologies, relying on HTML, CSS, and JavaScript to create the frontend interface of the system. For the backend, I employed PHP and integrated it with a MySQL database, ensuring the storage and retrieval of voting data while maintaining data integrity and security.

Given the sensitive nature of voting data and the prevalence of security threats in online applications, I took measures to secure the system against common risks. In particular, I addressed the potential threat of SQL injection attacks by implementing prepared statements in PHP, which effectively sanitized user input and prevented malicious SQL queries from compromising the system's integrity.

In summary, my final year project aimed to resolve issues associated with traditional voting methods in academic settings and, in doing so, create a platform that not only facilitated online elections but also did so securely and efficiently. This project provided a solution to the problems encountered in traditional voting systems within academic settings, with the potential to enhance the democratic processes and engagement levels of students.